fix: 디렉토리 컨밴션에 맞게 디렉토리 수정 (#11)
