+++ src/main/java/com/palette/infra/jwtTokenProvider/JwtAccessTokenInfo.java
@@ -1,14 +1,14 @@
-package com.palette.infrastructure.jwtTokenProvider;
+package com.palette.infra.jwtTokenProvider;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 @Component
-public class JwtRefreshTokenInfo implements JwtTokenInfo {
-    @Value("${security.jwt.refresh-token.secret-key}")
+public class JwtAccessTokenInfo implements JwtTokenInfo {
+    @Value("${security.jwt.access-token.secret-key}")
     private String secretKey;
 
-    @Value("${security.jwt.refresh-token.expire-length}")
+    @Value("${security.jwt.access-token.expire-length}")
     private Long validityInMilliseconds;
 
     @Override
@@ -28,6 +28,6 @@ public class JwtRefreshTokenInfo implements JwtTokenInfo {
 
     @Override
     public boolean supports(JwtTokenType jwtTokenType) {
-        return jwtTokenType.equals(JwtTokenType.REFRESH_TOKEN);
+        return jwtTokenType.equals(JwtTokenType.ACCESS_TOKEN);
     }
 }
