+++ src/main/java/com/palette/infra/jwtTokenProvider/JwtTokenType.java
@@ -1,4 +1,4 @@
-package com.palette.infrastructure.jwtTokenProvider;
+package com.palette.infra.jwtTokenProvider;
 
 public enum JwtTokenType {
     ACCESS_TOKEN, REFRESH_TOKEN;
